+++ src/utils/hisWritePayload.js
@@ -43,15 +43,23 @@ class HisWritePayload {
     /**
      * Add data to the hisWrite payload. If the unit is present on val attribute, it will be stripped.
      * @param id Entity id with Haystack prefix applied.
-     * @param data Data of the format of [{ts: String, val: String}] with the Haystack prefix applied.
+     * @param {[]} data Data of the format of [{ts: String, val: String}] with the Haystack prefix applied.
      */
     add(id, data) {
-        for (const {ts, val} of data) {
+        if (typeof id !== "string") {
+            throw new Error("Id must be a string");
+        } else if (id.substring(0, 2) !== "r:") {
+            throw new Error("Id must have Haystack reference type prefix applied");
+        }
+
+        for (const [i, {ts, val}] of data.entries()) {
             if (ts === undefined) {
                 throw new Error("Row in data missing 'ts' property");
             }
             else if (val === undefined) {
                 throw new Error("Row in data missing 'val' property");
+            } else if (typeof val !== "string" && typeof val !== "boolean") {
+                throw new Error(`'val' in row ${i} missing Haystack prefix`);
             }
 
             if (this.payload[ts] === undefined) {
@@ -59,7 +67,7 @@ class HisWritePayload {
             }
 
             let parsedVal = val;
-            if (val.substring(0, 2) === "n:" && val.includes(" ")) {
+            if (typeof val === "string" && val.substring(0, 2) === "n:" && val.includes(" ")) {
                 const index= val.indexOf(" ");
                 if (index !== -1) {
                     // val has unit applied, remove